requirements.txt
fastapi
uvicorn
supabase
python-dotenv
influxdb-client
